+++ CHANGELOG.md
@@ -34,10 +34,10 @@
 
 * First version. Released on an unsuspecting world.
 
-* Implement TypeID as specified at https://github.com/jetpack-io/typeid.
+* Implement `TypeID` as specified at https://github.com/jetpack-io/typeid.
 
 * Add unit tests.
 
-* Add type-level TypeID prefixes.
+* Add type-level `TypeID` prefixes.
 
 * Add `FromJSON` and `ToJSON` instances for `TypeID` and `KindID`.